move device config to it's own file
