+++ src/tests/unit/tests/electron/platform/android/fetch-device-config.test.ts
@@ -1,7 +1,8 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 import { AxiosResponse } from 'axios';
-import { createDeviceConfigFetcher, DeviceConfig, DeviceConfigFetcher } from 'electron/platform/android/device-config-fetcher';
+import { DeviceConfig } from 'electron/platform/android/device-config';
+import { createDeviceConfigFetcher, DeviceConfigFetcher } from 'electron/platform/android/device-config-fetcher';
 import { HttpGet } from 'electron/platform/android/fetch-scan-results';
 import { IMock, Mock } from 'typemoq';
